+++ src/lib.rs
@@ -10,3 +10,4 @@ pub mod scanner;
 pub mod scope;
 pub mod semantic;
 pub mod symbols;
+pub mod expression_constant_handler;